+++ client/src/Containers/Workspace/Tools/Awareness.js
@@ -1,0 +1,11 @@
+import React from "react";
+import classes from "./tools.css";
+const Awarness = props => {
+  let { awarenessDesc, awarenessIcon } = props;
+  return (
+    <div className={classes.ReferenceControls}>
+      <i className="far fa-eye" />
+      <div className={classes.AwarenessDesc}>{awarenessDesc}</div>
+    </div>
+  );
+};
