+++ client/src/Containers/Workspace/Tools/tools.css
@@ -1,6 +1,6 @@
 /* We should have made a component that chat currentMembers and this can
 all share. we are repeating styles in each of those copmnents  */
-:import("../../colors.css") {
+:import("../../../colors.css") {
   blue: blue;
   blueHover: blueHover;
 }
