+++ client/src/Containers/Workspace/index.js
@@ -4,7 +4,7 @@ export { default as DesmosGraph } from "./DesmosGraph";
 export { default as DesmosActivityGraph } from "./DesmosActivityGraph";
 export { default as Chat } from "./Chat";
 export { default as Tabs } from "./Tabs";
-export { default as Tools } from "./Tools";
+export { default as Tools } from "./Tools/Tools";
 export { default as RoomInfo } from "./RoomInfo";
 export { default as Workspace } from "./Workspace";
-export { default as ActivityTools } from "./ActivityTools";
+export { default as ActivityTools } from "./Tools/ActivityTools";
